requirements.txt
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0.1
websockets>=12.0
boto3>=1.34
orjson>=3.10
certifi>=2024.7.4
